Add cypress tests for the moderation archived state
