+++ cypress/support/helpers/dkan.js
@@ -184,3 +184,41 @@ export function generateDataDictionary(uuid) {
     }
   }
 }
+
+// create dataset with a given moderation state
+export function createDatasetWithModerationState(dataset_title, moderation_state) {
+  // Create a dataset.
+  cy.visit('/node/add/data')
+  cy.get('#edit-field-json-metadata-0-value-title')
+    .type(dataset_title, { force:true } )
+  cy.get('#edit-field-json-metadata-0-value-description')
+    .type('DKANTEST dataset description.', { force:true } )
+  cy.get('#edit-field-json-metadata-0-value-accesslevel')
+    .select('public', { force:true } )
+  cy.get('#edit-field-json-metadata-0-value-modified-date')
+    .type('2020-02-02', { force:true } )
+  // Fill select2 field for publisher.
+  cy.get('#edit-field-json-metadata-0-value-publisher-publisher-name + .select2')
+    .find('.select2-selection')
+    .click({ force:true })
+  cy.get('input[aria-controls="select2-edit-field-json-metadata-0-value-publisher-publisher-name-results"]')
+    .type('DKANTEST Publisher{enter}')
+  // End filling up publisher.
+  cy.get('#edit-field-json-metadata-0-value-contactpoint-contactpoint-fn')
+    .type('DKANTEST Contact Name', { force:true } )
+  cy.get('#edit-field-json-metadata-0-value-contactpoint-contactpoint-hasemail')
+    .type('dkantest@test.com', { force:true } )
+  // Fill select2 field for keyword.
+  cy.get('#edit-field-json-metadata-0-value-keyword-keyword-0 + .select2')
+    .find('.select2-selection')
+    .click({ force: true })
+  cy.get('input[aria-controls="select2-edit-field-json-metadata-0-value-keyword-keyword-0-results"]')
+    .type('open data{enter}')
+  cy.get('#edit-moderation-state-0-state')
+    .select(moderation_state, { force:true } )
+  // End filling up keyword.
+  cy.get('#edit-submit')
+    .click({ force:true })
+  cy.get('.messages--status')
+    .should('contain','has been created')
+}
